requirements.txt
blinker==1.8.2
chardet==5.2.0
click==8.1.7
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.3
openpyxl==3.1.5
pandas==2.2.3
pillow==11.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
reportlab==4.2.5
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.1.2
WTForms==3.2.1
